+++ src/main/java/use_case/loadsearch/LoadSearchHistoryInputData.java
@@ -1,4 +1,4 @@
-package use_case.autosave_search_history;
+package use_case.loadsearch;
 
 public class LoadSearchHistoryInputData {
