+++ src/main/java/use_case/loadsearch/SearchHistoryEntry.java
@@ -1,4 +1,4 @@
-package use_case.autosave_search_history;
+package use_case.loadsearch;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
