+++ src/main/java/use_case/savetopic/SaveTopicInputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.autosave_search_history;
+package use_case.savetopic;
 
 public interface SaveTopicInputBoundary {
     void execute(SaveTopicInputData inputData);
