+++ src/main/java/use_case/savetopic/SaveTopicInputData.java
@@ -1,4 +1,4 @@
-package use_case.autosave_search_history;
+package use_case.savetopic;
 
 import java.time.LocalDateTime;
